auth in contacts

+++ controllers/contactController.js
@@ -4,9 +4,9 @@ const asyncHandler = require('express-async-handler')
 const  Contact = require('../models/contactModel');
 //@desc Get all contacts
 //@route get /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async(req,res)=>{
-    const contact=await Contact.find();
+    const contact=await Contact.find({userid:req.user.id});
     res.status(200).json(contact)
 
 });
@@ -16,7 +16,7 @@ const getContacts = asyncHandler(async(req,res)=>{
 
 // @desc Create contact
 // @route post /api/contacts
-// @access public
+// @access private
 const createContact = asyncHandler(async(req,res)=>{
     console.log("the req body: ",req.body);
     const {name,phone}= req.body;
@@ -24,13 +24,13 @@ const createContact = asyncHandler(async(req,res)=>{
         res.status(400);
         throw new Error("All fields are mandatory")
     }
-    const contact = await Contact.create({name,phone})
+    const contact = await Contact.create({userid:req.user.id,name,phone})
     res.status(201).json(contact)
 });
 
 //@desc Get contact
 //@route get /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async(req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -42,13 +42,18 @@ const getContact = asyncHandler(async(req,res)=>{
 
 //@desc Update contact
 //@route put /api/contacts
-//@access public
+//@access private
 const updateContact = asyncHandler(async(req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact Not Found");
     }
+    if(contact.userid.toString()!==req.user.id){
+        res.status(403);
+        throw new Error("Users dont have permission to update other users contact")
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id,req.body,{new:true});
 
     res.status(200).json(updatedContact)
@@ -56,13 +61,18 @@ const updateContact = asyncHandler(async(req,res)=>{
 
 //@desc Delete contact
 //@route delete /api/contacts
-//@access public
+//@access private
 const deleteContact = asyncHandler(async(req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact Not Found");
     }
+    if(contact.userid.toString()!==req.user.id){
+        res.status(403);
+        throw new Error("Users dont have permission to delete other users contact")
+    }
+    
     await Contact.findByIdAndRemove(req.params.id);
     res.status(200).json(contact)
 });
